+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 而項目則基於 [Semantic Versioning](http://semver.org/lang/zh-TW/).
 
 ## [Unreleased]
+- 增加廣告版位新聞全網文末iframe `Nownews_${adType}_article_600x225_i_new2` @esbb48
 
 ## 1.0.27 - 2017-08-29
 ### Added